minor: rename document test file
